Moving js to ts and esbuild
