--- assets/js/element-index-2d7b2ba1.js
@@ -1,11 +1,0 @@
-const getElementIndex = (node) => {
-  let index = 0;
-  let currentSibling = node.previousElementSibling;
-  while (currentSibling) {
-    index += 1;
-    currentSibling = currentSibling.previousElementSibling;
-  }
-  return index;
-};
-
-export { getElementIndex as g };
