+++ config.yml
@@ -20,6 +20,3 @@ module:
     # this allows html/css/js to live side-by-side
     - source: layouts
       target: assets
-    # add js chunks to static
-    - source: assets/js
-      target: static/js
