--- jsconfig.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": [
-      "es5",
-      "dom"
-    ],
-    "target": "ES5",
-    "checkJs": true,
-    "noEmit": true,
-    "allowSyntheticDefaultImports": true,
-    "maxNodeModuleJsDepth": 0,
-    "skipLibCheck": true,
-    "module": "commonJS"
-  },
-  "include": [
-    "*.js",
-    "layouts/**/*.js",
-    "layouts/_default/*.js",
-    "globals.d.ts"
-  ]
-}
