+++ layouts/_default/page.html
@@ -6,7 +6,7 @@
 {{ $.Scratch.Add "css-critical" "partials/modules/content" }}
 {{ end }}
 {{ if .Params.buybutton }}
-{{ $.Scratch.Add "js-async" "page-buy-button.js" }}
+{{ $.Scratch.Add "ts" "_default/page-buy-button.ts" }}
 {{ end }}
 {{ end }}
