+++ layouts/_default/status.html
@@ -1,6 +1,6 @@
 {{ define "dependencies" }}
 {{ $.Scratch.Add "css-critical" "_default/status" }}
-{{ $.Scratch.Add "js-async" "status.js" }}
+{{ $.Scratch.Add "ts" "_default/status.js" }}
 {{ end }}
 
 {{ define "main" }}
