--- layouts/partials/js-async.html
@@ -1,15 +1,0 @@
-{{- range . | uniq -}}
-
-<!-- get main js -->
-{{- with resources.Get (printf "js/%s" .) }}
-{{- $js := . | fingerprint -}}
-<script async defer src="{{$js.RelPermalink}}" type="module"></script>
-{{- else -}}
-{{ warnf "Could not find js %s" . }}
-{{- end -}}
-
-{{- end -}}
-
-<script>
-  loaded = {{.}}
-</script>
