+++ layouts/products/single.html
@@ -6,7 +6,9 @@
 {{ $.Scratch.Add "css-critical" "partials/breadcrumb" }}
 {{ $.Scratch.Add "css-critical" "modal-critical" }}
 {{ $.Scratch.Add "css-async" "modal" }}
-{{ $.Scratch.Add "js-async" "product.js" }}
+{{ $.Scratch.Add "ts" "partials/product.js" }}
+{{ $.Scratch.Add "ts" "partials/elements/r-carousel.ts" }}
+{{ $.Scratch.Add "ts" "partials/elements/r-thumbnails.ts" }}
 {{ end }}
 
 {{ define "canonical" }}
