--- rollup-plugin-dynamic-import.js
@@ -1,11 +1,0 @@
-const plugin = () => ({
-  name: 'dynamic-import-polyfill',
-  renderDynamicImport() {
-    return {
-      left: '__import__(',
-      right: ')',
-    };
-  },
-});
-
-export default plugin;
