+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "lib": [
+      "ES2020",
+      "dom"
+    ],
+    "target": "ES5",
+    "noEmit": true,
+    "module": "ES2020"
+  },
+  "include": [
+    "layouts/**/*.ts",
+    "globals.d.ts"
+  ]
+}
